Optimize code & fixed bugs

+++ player/data.go
@@ -3,7 +3,7 @@ package player
 type PlayerData struct {
 	Name     string `json:"name"`
 	JoinDate string `json:"join_date"`
-	Coin     int64  `json:"coin"`
+	Coin     int    `json:"coin"`
 }
 
 func (p *PlayerData) GetName() string {
@@ -14,14 +14,14 @@ func (p *PlayerData) GetJoinDate() string {
 	return p.JoinDate
 }
 
-func (p *PlayerData) GetCoin() int64 {
+func (p *PlayerData) GetCoin() int {
 	return p.Coin
 }
 
-func (p *PlayerData) AddCoin(coin int64) {
+func (p *PlayerData) AddCoin(coin int) {
 	p.Coin += coin
 }
 
-func (p *PlayerData) ReduceCoin(coin int64) {
+func (p *PlayerData) ReduceCoin(coin int) {
 	p.Coin -= coin
 }
